% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_names.R
\name{name_text}
\alias{name_text}
\title{Matching your data with cshapes if you have country names.}
\usage{
name_text(countryname)
}
\arguments{
\item{countryname}{Country name as it appears in cshapes.}
}
\value{
The output will be the country war code, start and end date of the country in cshapes data.
}
\description{
\code{name_text} function to get country war codes, start and ending dates if you have country name.
}
