% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_names.R
\name{names_list}
\alias{names_list}
\title{Matching your data with cshapes - list of country names.}
\usage{
names_list(a)
}
\arguments{
\item{a}{it can be anything, or left empty.}
}
\value{
The output will be a list of all country name in cshapes.
}
\description{
\code{names_list} this function provides you with the list of all country names in cshapes.
}
