% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stadl_models.R
\name{ntsperr}
\alias{ntsperr}
\title{Estimates a Spatial error model (SEM) lag model.}
\usage{
ntsperr(lmobj, wm)
}
\arguments{
\item{lmobj}{An object created by the \code{lm} function.}

\item{wm}{neighbor spatial weight matrix created by the \code{make_ntspmat}.}
}
\value{
The output will be SAR model estimated.
}
\description{
\code{ntsperr} estimates a Spatial error model (SEM) model (spatialreg::errorsarlm)
using the nearest neighbor spatial weights matrix you previously created
and a linear regression model.
}
