% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stadl_models.R
\name{ntspreg}
\alias{ntspreg}
\title{Estimates a Spatial AutoRegressive (SAR) lag model.}
\usage{
ntspreg(lmobj, wm)
}
\arguments{
\item{lmobj}{An object created by the \code{lm} function.}

\item{wm}{neighbor spatial weight matrix created by the \code{make_ntspmat}.}
}
\value{
The output will be SAR model estimated.
}
\description{
\code{ntspreg} estimates a Spatial AutoRegressive (SAR) model (spatialreg::lagsarlm)
using the nearest neighbor spatial weights matrix you previously created
and a linear regression model.
}
