% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stadl_models.R
\name{ntspsac}
\alias{ntspsac}
\title{Estimates a Spatial Autocorrelation (SAC) lag model.}
\usage{
ntspsac(lmobj, wm)
}
\arguments{
\item{lmobj}{An object created by the \code{lm} function.}

\item{wm}{neighbor spatial weight matrix created by the \code{make_ntspmat}.}
}
\value{
The output will be SAR model estimated.
}
\description{
\code{ntspsac} estimates a Spatial Autocorrelation (SAC) model (spatialreg::sacsarlm)
using the nearest neighbor spatial weights matrix you previously created
and a linear regression model.
}
