% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_effects.R
\name{st_effects}
\alias{st_effects}
\title{Total, Direct, and Indirect Spatiotemporal Effects for the STADL (sy0,ty1) Model.}
\usage{
st_effects(spobj, wm, tlag, covariate)
}
\arguments{
\item{spobj}{An object created by the \code{ntspreg} function for a STADL (sy0,ty1) Model.}

\item{wm}{a list created by the \code{make_ntspmat} function that includes the spatial weights    #' matrix.}

\item{tlag}{Name of a variable that identifies the temporal lag of an outcome.}

\item{covariate}{Name of a variable to be shocked.}
}
\value{
The output will be a dataframe.
}
\description{
\code{st_effects} calculates both the short and long-run total, direct and indirect effects of shocks to sample units' covariates using the last cross-section of the data.
}
\examples{
 \dontrun{
fx <- st_effects(sar,wm,tlag,x)
}
}
