% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntspmat.R
\name{make_ntspmat}
\alias{make_ntspmat}
\title{NT Spatial Weights Matrix for Unbalanced TSCS Samples.}
\usage{
make_ntspmat(lmobj, ci, yi, k)
}
\arguments{
\item{lmobj}{An object created by the \code{lm} function.}

\item{ci}{Name of a variable that identifies countries.}

\item{yi}{Name of a variable that identifies years.}

\item{k}{A value that determines the number of nearest
neighbors.}
}
\value{
The output will be a list.
}
\description{
\code{make_ntspmat} generates a nearest neighbor spatial
weights matrix for an unbalanced TSCS sample used to
estimate a linear regression model.
}
\examples{
df <- data.frame(
country = factor(c("United Kingdom","United Kingdom","Ireland","Ireland",
"Netherlands","Netherlands", "Belgium","Belgium","France","France","Spain","Spain")),
year = c(2018, 2019, 2018, 2019, 2018, 2019, 2018, 2019, 2018, 2019, 2018, 2019),
y = rnorm(12), x = rnorm(12))
linmod <- lm(y~x,df)
w <- make_ntspmat(linmod,country,year,3)

}
