% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_names.R
\name{name_code}
\alias{name_code}
\title{Matching your data with cshapes if you have country war codes.}
\usage{
name_code(warcode)
}
\arguments{
\item{warcode}{Country war code (https://correlatesofwar.org/data-sets/cow-country-codes).}
}
\value{
The output will be the country name as it appears in cshapes, start and end date of the country in cshapes data.
}
\description{
\code{name_code} function to get country name in cshapes if you have war code.
It also provides you start and end date.
}
